+++ CHANGELOG.md
@@ -5,11 +5,13 @@
 
 ## All notable changes to this publication will be documented in this file.
 
-### 1.6.1 - 06/04/2020
+### 1.7 - 06/04/2020
 
 * Added an if statement that checks if the caches exist, and loads then despite the timer logic if they dont
 as they are required to display anything in the first place, also added TODOS.
 * Added for devs an auto-fetcher for external IP with phpdotenv if localhost is used to prevent geolocation to break
+* Added a checker for if cookies exist, if not- methods for getIP and fetch(location) will invoke and put user at their
+respective location, this cookie checker was added to prevent the geolocation API to be ratelimited
 
 ### 1.6 - 04/04/2020